chore(test): Migrate tests to Vitest and TypeScript
